// frontend/src/apisv2beta1/artifact/models/V2beta1ArtifactTask.ts
/* tslint:disable */
/* eslint-disable */
/**
 * backend/api/v2beta1/artifact.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { V2beta1IOType } from './V2beta1IOType';
import {
  V2beta1IOTypeFromJSON,
  V2beta1IOTypeFromJSONTyped,
  V2beta1IOTypeToJSON,
  V2beta1IOTypeToJSONTyped,
} from './V2beta1IOType';
import type { V2beta1IOProducer } from './V2beta1IOProducer';
import {
  V2beta1IOProducerFromJSON,
  V2beta1IOProducerFromJSONTyped,
  V2beta1IOProducerToJSON,
  V2beta1IOProducerToJSONTyped,
} from './V2beta1IOProducer';

/**
 *
 * @export
 * @interface V2beta1ArtifactTask
 */
export interface V2beta1ArtifactTask {
  /**
   * Output only. The unique server generated id of the ArtifactTask.
   * @type {string}
   * @memberof V2beta1ArtifactTask
   */
  readonly id?: string;
  /**
   *
   * @type {string}
   * @memberof V2beta1ArtifactTask
   */
  artifact_id?: string;
  /**
   *
   * @type {string}
   * @memberof V2beta1ArtifactTask
   */
  run_id?: string;
  /**
   *
   * @type {string}
   * @memberof V2beta1ArtifactTask
   */
  task_id?: string;
  /**
   *
   * @type {V2beta1IOType}
   * @memberof V2beta1ArtifactTask
   */
  type?: V2beta1IOType;
  /**
   *
   * @type {V2beta1IOProducer}
   * @memberof V2beta1ArtifactTask
   */
  producer?: V2beta1IOProducer;
  /**
   *
   * @type {string}
   * @memberof V2beta1ArtifactTask
   */
  key?: string;
}

/**
 * Check if a given object implements the V2beta1ArtifactTask interface.
 */
export function instanceOfV2beta1ArtifactTask(value: object): value is V2beta1ArtifactTask {
  return true;
}

export function V2beta1ArtifactTaskFromJSON(json: any): V2beta1ArtifactTask {
  return V2beta1ArtifactTaskFromJSONTyped(json, false);
}

export function V2beta1ArtifactTaskFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): V2beta1ArtifactTask {
  if (json == null) {
    return json;
  }
  return {
    id: json['id'] == null ? undefined : json['id'],
    artifact_id: json['artifact_id'] == null ? undefined : json['artifact_id'],
    run_id: json['run_id'] == null ? undefined : json['run_id'],
    task_id: json['task_id'] == null ? undefined : json['task_id'],
    type: json['type'] == null ? undefined : V2beta1IOTypeFromJSON(json['type']),
    producer: json['producer'] == null ? undefined : V2beta1IOProducerFromJSON(json['producer']),
    key: json['key'] == null ? undefined : json['key'],
  };
}

export function V2beta1ArtifactTaskToJSON(json: any): V2beta1ArtifactTask {
  return V2beta1ArtifactTaskToJSONTyped(json, false);
}

export function V2beta1ArtifactTaskToJSONTyped(
  value?: Omit<V2beta1ArtifactTask, 'id'> | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    artifact_id: value['artifact_id'],
    run_id: value['run_id'],
    task_id: value['task_id'],
    type: V2beta1IOTypeToJSON(value['type']),
    producer: V2beta1IOProducerToJSON(value['producer']),
    key: value['key'],
  };
}